requirements.txt
djangorestframework
django-cors-headers
python-dotenv
google-generativeai
